create SearchSuggestions component
